+++ src/Calculators/WonderCalculator.php
@@ -38,9 +38,7 @@ class WonderCalculator
      * @var float Constraints for RP gain formula
      */
     protected const TECH_MAX_REWARD = 4000;
-    protected const TECH_MIN_OFFNESE = 6000;
-    protected const TECH_OPTIMAL_SIZE = 590;
-    protected const TECH_SCALE_FACTOR = 1000;
+    protected const TECH_MIN_SIZE = 500;
 
     /**
      * Returns the wonder's power when being rebuilt.
@@ -139,23 +137,23 @@ class WonderCalculator
     /**
     * Calculates research point gain for a dominion
     *
+    * @param RoundWonder $wonder
     * @param Dominion $dominion
-    * @param float $offense
-    * @param float $defense
+    * @param float $offenseSent
     * @return float
     */
-    public function getTechGainForDominion(Dominion $dominion, float $offense, float $defense): float
+    public function getTechGainForDominion(RoundWonder $wonder, Dominion $dominion, float $offenseSent): float
     {
-        if ($offense < static::TECH_MIN_OFFNESE) {
+        $offenseRequired = $wonder->power * min(15, $dominion->round->daysInRound() - 3) / 100;
+        if ($offenseSent < $offenseRequired) {
             return 0;
         }
 
         $landCalculator = app(LandCalculator::class);
+        $totalLand = $landCalculator->getTotalLand($dominion);
 
-        $constant = static::TECH_MAX_REWARD * (static::TECH_SCALE_FACTOR ** 2);
-        $damageModifier = $offense / $defense;
-        $landModifier = abs($landCalculator->getTotalLand($dominion) - static::TECH_OPTIMAL_SIZE);
-        $techGain = $constant * $damageModifier / (($landModifier + static::TECH_SCALE_FACTOR) ** 2);
+        $scaleFactor = static::TECH_MIN_SIZE * 10;
+        $techGain = min(static::TECH_MAX_REWARD / (0.9 + ($totalLand / $scaleFactor)), static::TECH_MAX_REWARD);
 
         return $techGain;
     }